Refactor the ZFS snapshot script/library.
The ZFS library is now a module inside /zfs, and largely refactored to be no
longer specific to `shadow_copy` (i.e. functions like DeleteSnapshot, etc.)
It's also faster now as the `GetSnapshots` function correctly passes the
pool/file-system.

Time parsing in ZFS is now done off the snapshot directly, instead of
parsing strings which are snapshot name specific. This makes timestamps
available to all snapshot instances.

A bug in the default filter config is fixed. Previously, `__DEFAULT_FILTERS`
contained a class type instead of an instance for the last entry.